Rename package adapt to eventtypechooser
